Make function name match the method it implements
Function.checkArgs() used to have two versions, a non-debug and a debug
one. Since we’re not going to use the non-debug version because you can
never prove that a program contains no errors, the non-debug version was
removed along with the “if” statement which controlled which version would
be assigned to Function.checkArgs().

This commit removes the _debug suffix from the remaining function.