Change documentation comments’ style
Change file/class/member/method/function documentation comments to use the
special delimiters /** */ instead of ## for the first line. This hopefully
makes comments more readable, as it seems to be the case with the
similarly-styled doc comments in Python, and it also makes doc comments
more easily recognizable by a parser.